Added Oxygenna plugins
These are plugins that are requied by Oxygenna themes. There are a pain in the ass, because I only keep track of them in my local development WordPress instance. They are available on GitHub, but they have recursive submodules, and I can't figure out how to pull it all together to push to our server. So I add them manually. T_T